backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
qdrant-client==1.6.4
pydantic==2.5.2
pydantic-settings==2.1.0
sse-starlette==1.8.2
google-genai==0.2.1
PyMuPDF==1.23.8
python-multipart==0.0.6
tiktoken==0.8.0
